pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "scam-detection-env"
version = "1.0.0"
description = "AI scam detection & response training environment (OpenEnv)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.5,<3",
  "pyyaml>=6.0.1",
  "openai>=2.7.2",
  "openenv-core[core]>=0.2.3",
  "requests>=2.25.0",
]

[project.optional-dependencies]
cli = ["openenv-core[cli]>=0.2.3"]

[project.scripts]
# OpenEnv / HF: `uv run server` or validator expects this entry point
server = "server.app:main"
